+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### 1.4.11
 - Removed debug log configuration option, as it's no longer needed (raoulvdberge)
-- Fixed bug where storages that are removed remain visible (raoulvdberge)
-- Added support for External Storage on Interfaces and other Refined Storage blocks, so you can keep items in stock easier (raoulvdberge)
 - Removed "autocraft on redstone signal" option in the Crafter, use an External Storage in combination with an Interface with the Crafting Upgrade instead (raoulvdberge)
+- Fixed bug where storages that are removed remain visible (raoulvdberge)
+- Fixed bug where the GUI didn't close when a block is broken, causing a dupe bug with the Portable Grid (raoulvdberge)
+- Added support for External Storage on Interfaces and other Refined Storage blocks, so you can keep items in stock easier (raoulvdberge)
 
 ### 1.4.10
 - Improved performance of network scanning (raoulvdberge)